Removed the inversed_by correlations for multiple columns
The feature is nice, but relaistically would not be that helpful.  Most
of the data that would need to be loaded would need either a filter or
an order by which having the arrays of data would not have.  In addition,
since multiple entity types have references back and forth to one another,
using a serializer to output the data raised a circular dependency exception.